Revised Enemy AI
Hey I've remade the Enemy AI Control that included things that we need make the enemy be a bit challenging. I also re-added an enemy Prefab in the Unity project. I also got rid of the dash function in the player controller because we don't need that now. I also got rid of the y axis in the look rotation due to the fact it looks like floats towards the target.